try fix unit tests
