+++ custom_ops/xpu_ops/test/test_block_attn.py
@@ -16,6 +16,7 @@ import random
 
 import numpy as np
 import paddle
+from utils import init_tensor
 
 # block_attn_fused is deprecated and should be removed in the future
 from fastdeploy.model_executor.ops.xpu import (
@@ -38,59 +39,6 @@ def print_all_not_equal_elements_info(k, x, y):
     xy_max_abs_diff = paddle.max(paddle.abs(xy_diff))
     xy_min_abs_diff = paddle.min(paddle.abs(xy_diff))
     print(f"{k} mean diff: {xy_mean_diff}, max abs diff: {xy_max_abs_diff}, min abs diff: {xy_min_abs_diff}")
-
-
-def init_tensor(bsz, block_tables_shape):
-    encoder_batch_map = paddle.empty(bsz, dtype="int32")
-    decoder_batch_map = paddle.empty(bsz, dtype="int32")
-    encoder_batch_idx = paddle.empty(bsz, dtype="int32")
-    decoder_batch_idx = paddle.empty(bsz, dtype="int32")
-    encoder_seq_lod = paddle.empty(bsz + 1, dtype="int32")
-    decoder_seq_lod = paddle.empty(bsz + 1, dtype="int32")
-    encoder_kv_lod = paddle.empty(bsz + 1, dtype="int32")
-    prefix_len = paddle.empty(bsz, dtype="int32")
-    decoder_context_len = paddle.empty(bsz, dtype="int32")
-    decoder_context_len_cache = paddle.empty(bsz, dtype="int32")
-
-    prefix_block_tables = paddle.empty(block_tables_shape, dtype="int32")
-
-    encoder_batch_map_cpu = paddle.empty(bsz, dtype="int32", device="cpu")
-    decoder_batch_map_cpu = paddle.empty(bsz, dtype="int32", device="cpu")
-    encoder_batch_idx_cpu = paddle.empty(bsz, dtype="int32", device="cpu")
-    decoder_batch_idx_cpu = paddle.empty(bsz, dtype="int32", device="cpu")
-    encoder_seq_lod_cpu = paddle.empty(bsz + 1, dtype="int32", device="cpu")
-    decoder_seq_lod_cpu = paddle.empty(bsz + 1, dtype="int32", device="cpu")
-    encoder_kv_lod_cpu = paddle.empty(bsz + 1, dtype="int32", device="cpu")
-    prefix_len_cpu = paddle.empty(bsz, dtype="int32", device="cpu")
-    decoder_context_len_cpu = paddle.empty(bsz, dtype="int32", device="cpu")
-    decoder_context_len_cache_cpu = paddle.empty(bsz, dtype="int32", device="cpu")
-
-    len_info_cpu = paddle.empty(7, dtype="int32", device="cpu")
-
-    return (
-        encoder_batch_map,
-        decoder_batch_map,
-        encoder_batch_idx,
-        decoder_batch_idx,
-        encoder_seq_lod,
-        decoder_seq_lod,
-        encoder_kv_lod,
-        prefix_len,
-        decoder_context_len,
-        decoder_context_len_cache,
-        prefix_block_tables,
-        encoder_batch_map_cpu,
-        decoder_batch_map_cpu,
-        encoder_batch_idx_cpu,
-        decoder_batch_idx_cpu,
-        encoder_seq_lod_cpu,
-        decoder_seq_lod_cpu,
-        encoder_kv_lod_cpu,
-        prefix_len_cpu,
-        decoder_context_len_cpu,
-        decoder_context_len_cache_cpu,
-        len_info_cpu,
-    )
 
 
 def run_prefix_cache_block_attn(
